# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Swagger Test",
    "description": "Swagger autogen"
  },
  "host": "localhost:9090/api/v1",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Home"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/user/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Fetch all users",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "isActivated": {
          "type": "boolean",
          "example": true
        },
        "name": {
          "type": "string",
          "example": "String"
        },
        "userName": {
          "type": "string",
          "example": "String"
        }
      },
      "xml": {
        "name": "User"
      }
    }
  }
}